mod examples to follow mros2 v0.3.1
