+++ mros2.lib
@@ -1,1 +1,1 @@
-https://github.com/mROS-base/mros2#v0.2.3.1
+https://github.com/mROS-base/mros2#v0.3.1
