+++ workspace/pub_float32/app.cpp
@@ -16,18 +16,9 @@
 
 #include "mbed.h"
 #include "mros2.h"
-#include "std_msgs/msg/u_int16.hpp"
+#include "std_msgs/msg/float32.hpp"
 #include "EthernetInterface.h"
 
-mros2::Subscriber sub;
-mros2::Publisher pub;
-
-void userCallback(std_msgs::msg::UInt16 *msg)
-{
-  printf("subscribed msg: '%d'\r\n", msg->data);
-  printf("publishing msg: '%d'\r\n", msg->data);
-  pub.publish(*msg);
-}
 
 #define IP_ADDRESS ("192.168.11.2") /* IP address */
 #define SUBNET_MASK ("255.255.255.0") /* Subnet mask */
@@ -44,11 +35,33 @@ int main() {
   MROS2_DEBUG("mROS 2 initialization is completed\r\n");
 
   mros2::Node node = mros2::Node::create_node("mros2_node");
-  pub = node.create_publisher<std_msgs::msg::UInt16>("to_linux", 10);
-  sub = node.create_subscription<std_msgs::msg::UInt16>("to_stm", 10, userCallback);
-  std_msgs::msg::UInt16 msg;
-
+  mros2::Publisher pub = node.create_publisher<std_msgs::msg::Float32>("to_linux", 10);
   MROS2_INFO("ready to pub/sub message\r\n");
+
+  std_msgs::msg::Float32 msg;
+  auto publish_count = -0.5;
+  while (1)
+  {
+    msg.data = publish_count;
+    MROS2_INFO("publishing float msg!!");
+    pub.publish(msg);
+
+    if (0.0 >= msg.data)
+      printf("msg <= 0.0\r\n");
+    else if (0.0 < msg.data && msg.data < 0.5)
+      printf("0.0 < msg < 0.5\r\n");
+    else if (0.5 < msg.data && msg.data < 1.0)
+      printf("0.5 < msg < 1.0\r\n");
+    else
+      printf("msg >= 1.0\r\n");
+    /*
+     * you need to add `"target.printf_lib": "std"` into mbed_app.json
+    printf("publishing msg: '%f'\r\n", msg.data);
+     */
+
+    publish_count = publish_count + 0.1;
+    osDelay(1000);
+  }
   mros2::spin();
 
   return 0;
