+++ workspace/sub_uint16/app.cpp
@@ -16,33 +16,18 @@
 
 #include "mbed.h"
 #include "mros2.h"
-#include "std_msgs/msg/float32.hpp"
+#include "std_msgs/msg/u_int16.hpp"
 #include "EthernetInterface.h"
-
-mros2::Subscriber sub;
-mros2::Publisher pub;
-
-void userCallback(std_msgs::msg::Float32 *msg)
-{
-  if (0.0 >= msg->data) {
-    printf("msg <= 0.0\r\n");
-  } else if (0.0 < msg->data && msg->data < 0.5) {
-    printf("0.0 < msg < 0.5\r\n");
-  } else if (0.5 < msg->data && msg->data < 1.0) {
-    printf("0.5 < msg < 1.0\r\n");
-  } else {
-    printf("msg >= 1.0\r\n");
-  }
-  /* 
-  printf("subscribed msg: '%f'\r\n", msg->data);
-  printf("publishing msg: '%f'\r\n", msg->data);
-  */
-  pub.publish(*msg);
-}
 
 #define IP_ADDRESS ("192.168.11.2") /* IP address */
 #define SUBNET_MASK ("255.255.255.0") /* Subnet mask */
 #define DEFAULT_GATEWAY ("192.168.11.1") /* Default gateway */
+
+
+void userCallback(std_msgs::msg::UInt16 *msg)
+{
+  printf("subscribed msg: '%d'\r\n", msg->data);
+}
 
 int main() {
   EthernetInterface network;
@@ -55,9 +40,7 @@ int main() {
   MROS2_DEBUG("mROS 2 initialization is completed\r\n");
 
   mros2::Node node = mros2::Node::create_node("mros2_node");
-  pub = node.create_publisher<std_msgs::msg::Float32>("to_linux", 10);
-  sub = node.create_subscription<std_msgs::msg::Float32>("to_stm", 10, userCallback);
-  std_msgs::msg::Float32 msg;
+  mros2::Subscriber sub = node.create_subscription<std_msgs::msg::UInt16>("to_stm", 10, userCallback);
 
   MROS2_INFO("ready to pub/sub message\r\n");
   mros2::spin();
